Extract MessagesList component from ChatWindow
